Added code to Replit and added conversation_history variable
Added required Replit files but main.py is all you need as long as you have the required modules installed, this is just to make development a lot easier.

Added conversation_history which once the script is closed is completely deleted and next time you run the script it won't remember anything, but ChatGPT will remember anything you said previously in the conversation.