Added Dependency Check, if not found then installs with pip
Small feature added where it clears the console after installing/importing all the modules, to make the introduction look neat and clean.

At the start of the code it imports python built in modules so that if there is an error importing any modules it automatically downloads them and imports them, this way I can use PyInstaller to turn this script into an EXE file.