v1.0.9 BETA release
This is a BETA version and there is lots more to come, so far I've described this software as "basic" but I would like to push it beyond. I want this software to be simple yet rich of features, ease of use while making this as minimalistic as possible. No fancy HTML or CSS, I might make a web browser version of this in the future but so far this is me tinkering with AI using purely Python and console output.

Check out the previous commits for the features or look for "v1.0.9 BETA" in releases.